Move constraint wrapper into its own header
